// FrontEnd/src/app/models/user.model.ts
export class User {
  password: string;
  username: string;
  firstName: string;
  lastName: string;
  email: string;
  mobile: number;
  country: string;
  city: string;
  address: string;
  dateOfBirth: Date;
  lastLogin: Date;
  role: string;
  sex: string;
  employeeRole: string;
  userInformation: string;
}